requirements.txt
pyserial==2.7
RPi.GPIO==0.5.11
